Create function to find BMU

+++ src/utils/math_helper.cpp
@@ -8,6 +8,6 @@
 
 using namespace std;
 
-double euclidean_distance(Position pos1, Position pos2) {
+float euclidean_distance(Position pos1, Position pos2) {
     return sqrt(pow(pos1.x - pos2.x, 2) + pow(pos1.y - pos2.y, 2));
 }
